Pull out title rendering into util
