+++ packages/gestalt/src/ModuleExpandable.js
@@ -3,7 +3,7 @@ import React, { useState, type Node } from 'react';
 import Box from './Box.js';
 import Divider from './Divider.js';
 import ModuleExpandableItem from './ModuleExpandableItem.js';
-import { type ExpandableBaseProps } from './ModuleTypes.js';
+import { type ExpandableBaseProps } from './moduleTypes.js';
 
 export default function ModuleExpandable({
   id,
