+++ packages/gestalt/src/ModuleExpandableItem.js
@@ -4,7 +4,8 @@ import Box from './Box.js';
 import Icon from './Icon.js';
 import TapArea from './TapArea.js';
 import Text from './Text.js';
-import { type BaseProps, type ExpandableItemProps } from './ModuleTypes.js';
+import { type BaseProps, type ExpandableItemProps } from './moduleTypes.js';
+import { renderModuleTitle } from './moduleUtils.js';
 
 type Props = {|
   ...BaseProps,
@@ -24,17 +25,6 @@ export default function ModuleExpandableItem({
   type = 'info',
   children,
 }: Props): Node {
-  const EXPANDABLE_TYPE_ATTRIBUTES = {
-    info: {
-      icon,
-      color: 'darkGray',
-    },
-    error: {
-      icon: 'workflow-status-problem',
-      color: 'red',
-    },
-  };
-
   return (
     <>
       <TapArea
@@ -50,22 +40,7 @@ export default function ModuleExpandableItem({
         <Box padding={6} display="flex">
           <Box display="flex" flex="grow" marginEnd={6} alignItems="baseline">
             <Box column={isCollapsed && summary ? 6 : 12} display="flex">
-              {EXPANDABLE_TYPE_ATTRIBUTES[type].icon && (
-                <Box marginEnd={2}>
-                  <Icon
-                    icon={EXPANDABLE_TYPE_ATTRIBUTES[type].icon}
-                    accessibilityLabel={iconAccessibilityLabel || ''}
-                    color={EXPANDABLE_TYPE_ATTRIBUTES[type].color}
-                  />
-                </Box>
-              )}
-              <Text
-                weight="bold"
-                truncate
-                color={EXPANDABLE_TYPE_ATTRIBUTES[type].color}
-              >
-                {title}
-              </Text>
+              {renderModuleTitle(type, icon, iconAccessibilityLabel, title)}
             </Box>
             {summary && isCollapsed && (
               <Box column={6} marginStart={6}>
